requirements.txt
fastapi
uvicorn
langchain
langchain_community
langchain-openai
langchain_chroma
python-dotenv
azure-devops
streamlit
requests
pypdf